Add some aliases
Damn I can't be bothered to write proper commit messages after I just lost half
an hour worth of them. I described the reasoning of the change, the  intended
workflow, spent a lot of time rewriting it to be well readable and even made
footnotes with references. Then when I :wq I get

hint: Waiting
for your editor to close the file... error: There was a problem with the editor
'vi'.
Please supply the message using either -m or -F option.

in return. I suppose that is what I get for using VI, thanks Saint IGNUcius. To
be fair Emacs has a godawful backup system that pollutes every god damn
directory with backup files. Right now having that in place wouldn't be so bad.
Well, well. This is not my first time losing data and certainly not the last. I
thought vim + git wouldn't fail on me like that, but what do you know. I should
start employing the good old regular copy paste to TextEdit.app for longer git
messages too.

If anyone by any chance is reading this I thought it'd be worth mentioning that
the message I spent half an hour wasn't for this commit. I just felt like
ranting and suddenly had a textfield to do it in.

Thanks for reading my blog